Extract light page view helpers
